requirements.txt
Pillow~=10.2.0
exifread~=3.0.0